re #3725. Clearing warnings

+++ Code/Mantid/Framework/CurveFitting/test/CompositeFunctionTest.h
@@ -219,7 +219,7 @@ public:
     TS_ASSERT_THROWS(mfun->setMatrixWorkspace(ws,7,12,3),std::range_error);
     
     mfun->setMatrixWorkspace(ws,0);
-    mfun->testInit(ws,0,0,ws->blocksize()-1);
+    mfun->testInit(ws,0,0,static_cast<int>(ws->blocksize())-1);
 
     mfun->setMatrixWorkspace(ws,0,1.0,3.3);
     mfun->testInit(ws,0,10,32);
